cambiados bigdecimal para precision

+++ src/main/java/com/cryptocurrencies/api/domain/model/TransactionHistory.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -14,8 +15,8 @@ public class TransactionHistory
     private Long id;
     private Long idUser;
     private String symbol;
-    private double actualPrice;
-    private double amount;
+    private BigDecimal actualPrice;
+    private BigDecimal amount;
     private char opType;
     private LocalDateTime transactionTime;
 }
